+++ tests/mcmc/test_sampling.py
@@ -184,7 +184,7 @@ class LinearRegressionTest(chex.TestCase):
             blackjax_mclmc_sampler_params,
             params_history,
             final_da,
-        ) = blackjax.adaptation.mclmc_adaptation.adjusted_mclmc_find_L_and_step_size(
+        ) = blackjax.adjusted_mclmc_find_L_and_step_size(
             mclmc_kernel=kernel,
             num_steps=num_steps,
             state=initial_state,
